refac. FAQ
